feat: added service to event and address event

+++ src/main/java/com/mundim/ticketbackendspringboot/repository/EventRepository.java
@@ -1,10 +1,9 @@
 package com.mundim.ticketbackendspringboot.repository;
 
-import com.mundim.ticketbackendspringboot.entity.Account;
 import com.mundim.ticketbackendspringboot.entity.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> {
 }
